+++ lib/vagrant-mutagen-utilize/command.rb
