+++ lib/vagrant-mutagen-utilize/config.rb
